backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pytesseract==0.3.10
pdf2image==1.17.0
chromadb==0.4.22
anthropic==0.18.1
python-dotenv==1.0.1
pillow==10.2.0
numpy==1.26.4
pytest==8.0.1
httpx==0.26.0 